+++ src/main/java/racingcar/domain/dto/CarMovementDto.java
@@ -1,6 +1,6 @@
 package racingcar.domain.dto;
 
-import racingcar.domain.game.CarName;
+import racingcar.domain.game.car.CarName;
 
 public record CarMovementDto(CarName carName, int position) {
 }
