+++ src/main/java/racingcar/domain/game/move/RandomMoveStrategy.java
@@ -1,4 +1,7 @@
-package racingcar.util;
+package racingcar.domain.game.move;
+
+import racingcar.util.MoveStrategy;
+import racingcar.util.NumberGenerator;
 
 public class RandomMoveStrategy implements MoveStrategy {
     private static final int THRESHOLD = 4;
