+++ src/test/java/racingcar/converter/InputConverterTest.java
@@ -7,7 +7,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import racingcar.domain.game.Cars;
+import racingcar.domain.game.car.Cars;
 
 class InputConverterTest {
